ref(slack): convert shield to boostrap
